+++ src/Git_Bulk_Updater/Updater.php
@@ -22,6 +22,14 @@ if ( ! defined( 'WPINC' ) ) {
  * Class Updater
  */
 class Updater {
+	use Webhooks;
+
+	/**
+	 * Constructor.
+	 */
+	public function __construct() {
+		$this->init( GIT_BULK_UPDATER_DIR );
+	}
 
 	/**
 	 * Use wp_remote_get() on individual webhooks.
@@ -33,11 +41,24 @@ class Updater {
 			if ( ! check_admin_referer( 'git_bulk_updater_nonce', 'git_bulk_updater_nonce' ) ) {
 				return;
 			}
-			$site     = str_replace( '_', '.', array_search( 'Update', $_POST, true ) );
-			$webhooks = ( new Webhooks() )->run( GIT_BULK_UPDATER_DIR );
-			$sites    = ( new Webhooks() )->parse_webhooks( $webhooks );
-			foreach ( $sites[ $site ]['all'] as $webhook ) {
-				wp_remote_get( $webhook );
+			$webhooks = [];
+			$update   = array_search( 'Update', $_POST, true );
+			$type     = false !== strpos( $update, 'plugin_' ) ? 'plugin' : null;
+			$type     = false !== strpos( $update, 'theme_' ) ? 'theme' : $type;
+			$update   = 'plugin' === $type ? str_replace( 'plugin_', '', $update ) : $update;
+			$update   = 'theme' === $type ? str_replace( 'theme_', '', $update ) : $update;
+			$site     = null === $type ? str_replace( '_', '.', $update ) : null;
+			$webhooks = null !== $site ? $this->all_webhooks[ $site ] : $webhooks;
+			if ( null === $site ) {
+				$repos = $this->repos[ $update ];
+				unset( $repos['sites'] );
+				foreach ( $repos as $repo ) {
+					$webhooks[] = $repo['url'];
+				}
+			}
+
+			foreach ( $webhooks as $webhook ) {
+				$response = wp_remote_get( $webhook );
 			}
 			( new Actions() )->redirect();
 		}