+++ CHANGES.md
@@ -1,3 +1,4 @@
 #### [unreleased]
 * initial pass
+* add rows for both update entire site or update all sites with specific repository
 * add ability to use multiple JSON files.